@@ -9,18 +9,18 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-picker/picker": "2.4.2",
+    "@react-native-picker/picker": "2.6.1",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
-    "expo": "~46.0.16",
-    "expo-linear-gradient": "~11.4.0",
-    "expo-status-bar": "~1.4.0",
+    "expo": "^50.0.13",
+    "expo-linear-gradient": "~12.7.2",
+    "expo-status-bar": "~1.11.1",
     "prop-types": "^15.8.1",
-    "react": "18.0.0",
-    "react-native": "0.69.9",
-    "react-native-gesture-handler": "~2.5.0",
-    "react-native-safe-area-context": "4.3.1",
-    "react-native-screens": "~3.15.0",
+    "react": "18.2.0",
+    "react-native": "0.73.4",
+    "react-native-gesture-handler": "~2.14.0",
+    "react-native-safe-area-context": "4.8.2",
+    "react-native-screens": "~3.29.0",
     "react-native-table-component": "^1.2.0",
     "react-navigation": "^4.4.4"
   },
