@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import { LinearGradient } from "expo-linear-gradient";
 import {
   Text,
@@ -12,140 +12,111 @@ import {
 import { createPet } from "../api";
 import { Picker } from "@react-native-picker/picker";
 
-class CreateScreen extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      id: null,
-      name: null,
-      status: "available",
-    };
-  }
+const CreateScreen = ({ navigation }) => {
+  const [petData, setPetData] = useState({
+    id: null,
+    name: null,
+    status: "available",
+  });
 
-  createAlertSuccess() {
+  const createAlertSuccess = () => {
     Alert.alert("Alerta Succes", "Intrarea a fost adaugata cu succes", [
-      { text: "OK", onPress: () => this.props.navigation.navigate("list", {}) },
+      { text: "OK", onPress: () => navigation.navigate("list") },
     ]);
-  }
+  };
 
-  createAlertValidate() {
+  const createAlertValidate = () => {
     Alert.alert(
       "Alerta Validare",
       "Datele de intrare sunt incorecte. Va rugam verificati.",
-      [
-        {
-          text: "OK",
-        },
-      ]
+      [{ text: "OK" }]
     );
-  }
+  };
 
-  createAlertError() {
+  const createAlertError = () => {
     Alert.alert(
       "Alerta Eroare",
       "Datele nu au fost create. Va rugam incercati mai tarziu in cazul in care nu merge serverul.",
-      [
-        {
-          text: "OK",
-        },
-      ]
+      [{ text: "OK" }]
     );
-  }
+  };
 
-  addPet() {
+  const addPet = () => {
     // Validate form inputs
-    if (
-      this.state.id === null ||
-      this.state.name === null ||
-      this.state.status === null ||
-      this.state.id.length === 0 ||
-      this.state.name.length === 0 ||
-      isNaN(this.state.id) ||
-      this.state.status.length === 0
-    ) {
-      this.createAlertValidate();
+    const { id, name, status } = petData;
+    if (!id || !name || !status || isNaN(id) || !status.length) {
+      createAlertValidate();
     } else {
-      createPet(this.state.id, this.state.name, this.state.status).then(
-        (response) => {
-          if (response) {
-            this.createAlertSuccess();
-          } else {
-            this.createAlertError();
-          }
+      createPet(id, name, status).then((response) => {
+        if (response) {
+          createAlertSuccess();
+        } else {
+          createAlertError();
         }
-      );
+      });
     }
-  }
+  };
 
-  setId(newText) {
-    const state = this.state;
-    state.id = newText;
-    this.setState(state);
-  }
-  setName(newText) {
-    const state = this.state;
-    state.name = newText;
-    this.setState(state);
-  }
-  setStatus(newText) {
-    const state = this.state;
-    state.status = newText;
-    this.setState(state);
-  }
-
-  render() {
-    return (
-      <LinearGradient colors={["#ccffcc", "white"]} style={styles.container}>
-        <ImageBackground
-          source={require("../assets/marks.png")}
-          resizeMode="cover"
-          style={styles.container}
-          imageStyle={styles.backgroundimage}
-        >
-          <View style={styles.Intupadding}>
-            <Text>Id:</Text>
-            <TextInput
-              style={styles.input}
-              placeholder="Please enter new id"
-              onChangeText={(newText) => this.setId(newText)}
-              keyboardType="number-pad"
-              autoCapitalize="none"
-              autoCorrect={false}
-            />
-            <Text>Nume:</Text>
-            <TextInput
-              style={styles.input}
-              placeholder="Please enter new name"
-              onChangeText={(newText) => this.setName(newText)}
-              autoCapitalize="none"
-              autoCorrect={false}
-            />
-            <Text>Status:</Text>
-            <Picker
-              selectedValue={this.state.status}
-              onValueChange={(currentStatus) => this.setStatus(currentStatus)}
-            >
-              <Picker.Item label="Available" value="available" />
-              <Picker.Item label="Pending" value="pending" />
-              <Picker.Item label="Sold" value="sold" />
-            </Picker>
-            <View style={styles.btn}>
-              <TouchableOpacity onPress={() => this.addPet()}>
-                <Text style={styles.btnText}>Adaugare</Text>
-              </TouchableOpacity>
-            </View>
+  return (
+    <LinearGradient colors={["#ccffcc", "white"]} style={styles.container}>
+      <ImageBackground
+        source={require("../assets/marks.png")}
+        resizeMode="cover"
+        style={styles.container}
+        imageStyle={styles.backgroundimage}
+      >
+        <View style={styles.Inputpadding}>
+          <Text>Id:</Text>
+          <TextInput
+            style={styles.input}
+            placeholder="Please enter new id"
+            onChangeText={(newText) =>
+              setPetData((prevState) => ({ ...prevState, id: newText }))
+            }
+            keyboardType="number-pad"
+            autoCapitalize="none"
+            autoCorrect={false}
+          />
+          <Text>Nume:</Text>
+          <TextInput
+            style={styles.input}
+            placeholder="Please enter new name"
+            onChangeText={(newText) =>
+              setPetData((prevState) => ({ ...prevState, name: newText }))
+            }
+            autoCapitalize="none"
+            autoCorrect={false}
+          />
+          <Text>Status:</Text>
+          <Picker
+            selectedValue={petData.status}
+            onValueChange={(currentStatus) =>
+              setPetData((prevState) => ({
+                ...prevState,
+                status: currentStatus,
+              }))
+            }
+          >
+            <Picker.Item label="Available" value="available" />
+            <Picker.Item label="Pending" value="pending" />
+            <Picker.Item label="Sold" value="sold" />
+          </Picker>
+          <View style={styles.btn}>
+            <TouchableOpacity onPress={addPet}>
+              <Text style={styles.btnText}>Adaugare</Text>
+            </TouchableOpacity>
           </View>
-        </ImageBackground>
-      </LinearGradient>
-    );
-  }
-}
+        </View>
+      </ImageBackground>
+    </LinearGradient>
+  );
+};
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
-  Intupadding: {
+  Inputpadding: {
     padding: 16,
   },
   backgroundimage: {
