@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState, useEffect } from "react";
 import {
   StyleSheet,
   View,
@@ -8,206 +8,171 @@ import {
   Modal,
   Pressable,
   StatusBar,
+  ScrollView,
 } from "react-native";
-import { ScrollView } from "react-native-gesture-handler";
 import { Table, TableWrapper, Row, Cell } from "react-native-table-component";
 import { withNavigation } from "react-navigation";
 import { findPets, deletePet } from "../api";
 import { Picker } from "@react-native-picker/picker";
 
-class ListScreen extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      tableHead: ["Index", "Nume", "Status", "Actiuni"],
-      tableData: [],
-      status: "available",
-      modalVisible: false,
-      idToRemove: null,
-    };
-    this.getPetsByStatus(this.state.status);
-  }
+const ListScreen = ({ navigation }) => {
+  const [tableData, setTableData] = useState([]);
+  const [status, setStatus] = useState("available");
+  const [modalVisible, setModalVisible] = useState(false);
+  const [idToRemove, setIdToRemove] = useState(null);
 
-  _Create() {
-    this.props.navigation.navigate("create");
-  }
+  useEffect(() => {
+    getPetsByStatus(status);
+  }, [status]);
 
-  getPetsByStatus(status) {
+  const _Create = () => {
+    navigation.navigate("create");
+  };
+
+  const getPetsByStatus = (status) => {
     findPets(status).then((response) => {
-      const state = this.state;
-      state.tableData = response.splice(0, 50);
-      this.setState(state);
+      setTableData(response.slice(0, 50));
     });
-  }
+  };
 
-  setStatus(currentStatus) {
-    const state = this.state;
-    state.status = currentStatus;
-    this.setState(state);
-    this.getPetsByStatus(this.state.status);
-  }
+  const _GoToEditScreen = (index) => {
+    navigation.navigate("edit", { id: tableData[index].id });
+  };
 
-  _GoToEditScreen(index) {
-    this.props.navigation.navigate("edit", {
-      id: this.state.tableData[index].id,
-    });
-  }
+  const _GoToViewScreen = (index) => {
+    navigation.navigate("view", { id: tableData[index].id });
+  };
 
-  _GoToViewScreen(index) {
-    this.props.navigation.navigate("view", {
-      id: this.state.tableData[index].id,
-    });
-  }
+  const _Delete = (index) => {
+    setIdToRemove(tableData[index].id);
+    setModalVisible(true);
+  };
 
-  _Delete(index) {
-    const state = this.state;
-    state.idToRemove = this.state.tableData[index].id;
-    this.setState(state);
-    this.setModalVisible(true);
-  }
-
-  removePet() {
-    deletePet(this.state.idToRemove).then((response) => {
-      this.setModalVisible(false);
+  const removePet = () => {
+    deletePet(idToRemove).then((response) => {
+      setModalVisible(false);
       if (response) {
-        this.createAlertSuccess();
-        this.getPetsByStatus(this.state.status);
+        createAlertSuccess();
+        getPetsByStatus(status);
       } else {
-        this.createAlertError();
+        createAlertError();
       }
     });
-  }
+  };
 
-  createAlertSuccess() {
+  const createAlertSuccess = () => {
     Alert.alert("Alerta Succes", "Intrarea a fost stearsa cu succes", [
-      { text: "OK", onPress: () => this.props.navigation.navigate("list", {}) },
+      { text: "OK", onPress: () => navigation.navigate("list", {}) },
     ]);
-  }
+  };
 
-  createAlertError() {
+  const createAlertError = () => {
     Alert.alert(
       "Alerta Eroare",
       "Datele nu au fost sterse. Va rugam incercati mai tarziu in cazul in care nu merge serverul.",
-      [
-        {
-          text: "OK",
-        },
-      ]
+      [{ text: "OK" }]
     );
-  }
-
-  setModalVisible = (visible) => {
-    this.setState({ modalVisible: visible });
   };
 
-  render() {
-    const state = this.state;
-    // View , Edit, Delete buttons
-    const element = (data, index) => (
-      <View>
-        <TouchableOpacity onPress={() => this._GoToViewScreen(index)}>
-          <View style={styles.btnView}>
-            <Text style={styles.btnTextTable}>vizualizare</Text>
-          </View>
-        </TouchableOpacity>
-
-        <TouchableOpacity onPress={() => this._GoToEditScreen(index)}>
-          <View style={styles.btnEdit}>
-            <Text style={styles.btnTextTable}>editare</Text>
-          </View>
-        </TouchableOpacity>
-
-        <TouchableOpacity onPress={() => this._Delete(index)}>
-          <View style={styles.btnDelete}>
-            <Text style={styles.btnTextTable}>stergere</Text>
-          </View>
-        </TouchableOpacity>
-      </View>
-    );
-
-    return (
-      <>
-        <StatusBar style="white" />
-        <View style={styles.container}>
-          <View style={styles.btn} android_ripple={{ color: "green" }}>
-            <TouchableOpacity onPress={() => this._Create()}>
-              <Text style={styles.btnText}>Adaugare</Text>
-            </TouchableOpacity>
-          </View>
-          <Picker
-            style={styles.pick}
-            selectedValue={this.state.status}
-            onValueChange={(currentStatus) => this.setStatus(currentStatus)}
-          >
-            <Picker.Item label="Available" value="available" />
-            <Picker.Item label="Pending" value="pending" />
-            <Picker.Item label="Sold" value="sold" />
-          </Picker>
-          <ScrollView style={styles.scrollbar}>
-            <Table borderStyle={{ borderColor: "transparent", borderWidth: 1 }}>
-              <Row
-                data={state.tableHead}
-                style={styles.head}
-                textStyle={styles.text}
-              />
-              {state.tableData.map((rowData, index) => (
-                <TableWrapper key={index} style={styles.row}>
-                  {
-                    <>
-                      <Cell textStyle={styles.text} data={index + 1} />
-                      <Cell textStyle={styles.text} data={rowData.name} />
-                      <Cell textStyle={styles.text} data={rowData.status} />
-                      <Cell
-                        textStyle={styles.text}
-                        data={element(rowData, index)}
-                      />
-                    </>
-                  }
-                </TableWrapper>
-              ))}
-            </Table>
-          </ScrollView>
-
-          <Modal
-            animationType="slide"
-            transparent={true}
-            visible={this.state.modalVisible}
-            onRequestClose={() => {
-              Alert.alert("Modal has been closed.");
-              this.setModalVisible(!this.state.modalVisible);
-            }}
-          >
-            <View style={styles.centeredView}>
-              <View style={styles.modalView}>
-                <Text style={styles.modalText}>
-                  Sunteti sigur ca vreti sa stergeti?
-                </Text>
-                <Pressable
-                  style={[styles.button, styles.buttonClose]}
-                  onPress={() => this.removePet()}
-                >
-                  <Text style={styles.textStyle}>Da</Text>
-                </Pressable>
-                <Pressable
-                  style={[styles.button, styles.buttonClose]}
-                  onPress={() => this.setModalVisible(!this.state.modalVisible)}
-                >
-                  <Text style={styles.textStyle}>Nu</Text>
-                </Pressable>
-              </View>
-            </View>
-          </Modal>
+  return (
+    <>
+      <StatusBar style="white" />
+      <View style={styles.container}>
+        <View style={styles.btn} android_ripple={{ color: "green" }}>
+          <TouchableOpacity onPress={_Create}>
+            <Text style={styles.btnText}>Adaugare</Text>
+          </TouchableOpacity>
         </View>
-      </>
-    );
-  }
-}
+        <Picker
+          style={styles.pick}
+          selectedValue={status}
+          onValueChange={(currentStatus) => setStatus(currentStatus)}
+        >
+          <Picker.Item label="Available" value="available" />
+          <Picker.Item label="Pending" value="pending" />
+          <Picker.Item label="Sold" value="sold" />
+        </Picker>
+        <ScrollView style={styles.scrollbar}>
+          <Table borderStyle={{ borderColor: "transparent", borderWidth: 1 }}>
+            <Row
+              data={["Index", "Nume", "Status", "Actiuni"]}
+              style={styles.head}
+              textStyle={styles.text}
+            />
+            {tableData.map((rowData, index) => (
+              <TableWrapper key={index} style={styles.row}>
+                <Cell textStyle={styles.text} data={index + 1} />
+                <Cell textStyle={styles.text} data={rowData.name} />
+                <Cell textStyle={styles.text} data={rowData.status} />
+                <Cell
+                  textStyle={styles.text}
+                  data={
+                    <View>
+                      <TouchableOpacity onPress={() => _GoToViewScreen(index)}>
+                        <View style={styles.btnView}>
+                          <Text style={styles.btnTextTable}>vizualizare</Text>
+                        </View>
+                      </TouchableOpacity>
+
+                      <TouchableOpacity onPress={() => _GoToEditScreen(index)}>
+                        <View style={styles.btnEdit}>
+                          <Text style={styles.btnTextTable}>editare</Text>
+                        </View>
+                      </TouchableOpacity>
+
+                      <TouchableOpacity onPress={() => _Delete(index)}>
+                        <View style={styles.btnDelete}>
+                          <Text style={styles.btnTextTable}>stergere</Text>
+                        </View>
+                      </TouchableOpacity>
+                    </View>
+                  }
+                />
+              </TableWrapper>
+            ))}
+          </Table>
+        </ScrollView>
+
+        <Modal
+          animationType="slide"
+          transparent={true}
+          visible={modalVisible}
+          onRequestClose={() => {
+            Alert.alert("Modal has been closed.");
+            setModalVisible(false);
+          }}
+        >
+          <View style={styles.centeredView}>
+            <View style={styles.modalView}>
+              <Text style={styles.modalText}>
+                Sunteti sigur ca vreti sa stergeti?
+              </Text>
+              <Pressable
+                style={[styles.button, styles.buttonClose]}
+                onPress={removePet}
+              >
+                <Text style={styles.textStyle}>Da</Text>
+              </Pressable>
+              <Pressable
+                style={[styles.button, styles.buttonClose]}
+                onPress={() => setModalVisible(false)}
+              >
+                <Text style={styles.textStyle}>Nu</Text>
+              </Pressable>
+            </View>
+          </View>
+        </Modal>
+      </View>
+    </>
+  );
+};
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     padding: 16,
     backgroundColor: "#ccffcc",
+    paddingBottom: 50,
   },
   head: {
     height: 40,
@@ -218,9 +183,6 @@ const styles = StyleSheet.create({
     margin: 6,
     color: "black",
     textAlign: "center",
-  },
-  img: {
-    alignItems: "center",
   },
   row: {
     flexDirection: "row",
@@ -287,9 +249,6 @@ const styles = StyleSheet.create({
     padding: 10,
     elevation: 2,
     margin: 2,
-  },
-  buttonOpen: {
-    backgroundColor: "#F194FF",
   },
   buttonClose: {
     backgroundColor: "#2196F3",
